pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gigachat-synthetic-news"
version = "0.1.0"
description = "Synthetic financial news generation pipelines powered by GigaChat."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain",
    "numpy",
    "openpyxl",
    "pandas",
    "PyYAML",
    "tqdm",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
